requirements.txt
opencv-python-headless
pandas
numpy
matplotlib
scikit-learn
tensorflow-cpu
tqdm